feat : 게시글 삭제 방식 soft delete로 변경
기존 hard delete에서 삭제 방식을 soft delete로 변경.
다른 사람이 작성한 댓글이 남아있으면 삭제할 수 없다.

+++ src/main/java/codesquad/springcafe/controller/ArticleController.java
@@ -117,6 +117,10 @@ public class ArticleController {
     public String deleteArticle(@PathVariable("articleId") String articleId, HttpSession session) {
         User writer = (User) session.getAttribute("sessionUser");
         articleValidator.validWriter(writer, articleId);
+        if (replyService.existReply(articleId) && replyService.existOtherReply(articleId, writer.getUserId())) {
+            throw new InvalidAccessException("게시글을 삭제 불가능합니다.");
+        }
+        replyService.deleteAllReply(articleId);
         articleService.delete(articleId);
         return "redirect:/";
     }
